Moved NPC to its own class file to facilitate adding more NPC types. Made movement more fluid (worked around key repeat rate delay, and allowed players to continue moving in a non-colliding direction even if colliding in another). Added a sprint button (shift)
